+++ src/HDInsight/HDInsight/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 
 ## Upcoming Release
+* Fixed secrets exposure in example documentation.
 
 ## Version 6.1.0
 * Added new feature: Enable secure channels while creating a new cluster.