+++ src/HDInsight/HDInsight/ChangeLog.md
@@ -20,7 +20,7 @@
 
 ## Upcoming Release
 * Added new feature: Enable adding public IP tags to clusters. 
-* Added commands for manage Azure Montior Agent
+* Added commands for manage Azure Monitor Agent
     - Command `Get-AzHDInsightAzureMonitorAgent` to get the Azure Monitor Agent status of HDInsight cluster.
     - Command `Enable-AzHDInsightAzureMonitorAgent` to enable the Azure Monitor Agent in HDInsight cluster.
     - Command `Disable-AzHDInsightAzureMonitorAgent` to disable the Azure Monitor Agent in HDInsight cluster.